Install newer user-event as their doc mentioned
@testing-library/user-event@^14.0.0-beta provides a user interaction session management
via `userEvent.setup()`. It is useful because we can write tests as if an user actually
controls a components in their browser, and is recommended in the doc.

Also I remove @testing-library/dom from package.json because it is automatically
installed with @testing-library/react, and we should let the former's version be
resolved by the latter.
cf: https://testing-library.com/docs/user-event/install
Note that this doesn't mean the former package gets uninstalled from the project.